requirement.txt
aniso8601==9.0.1
anyio==3.6.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
docopt==0.6.2
et-xmlfile==1.1.0
fastapi==0.92.0
Flask==2.2.2
Flask-Compress==1.13
Flask-RESTful==0.3.9
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
openpyxl==3.1.1
pandas==1.3.5
pipreqs==0.4.11
plotly==5.13.0
psycopg2==2.9.5
pydantic==1.10.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tenacity==8.2.0
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.12.1